cedrus: Add a buildroot external folder
Rely on the BR2_EXTERNAL setup to create a minimal package for our libva
backend.

We also create a defconfig to create the toolchain with the kernel headers
patched to include the request API, based on the glibc and a gcc7.

This will be useful to automate our build, in order to provide some CI.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>